backend/app/requirements.txt

fastapi
uvicorn[standard]
requests
python-dotenv
faiss-cpu
sentence-transformers
pydantic
